double fix

+++ Assets/Scripts/Player.cs
@@ -5,14 +5,14 @@ public class Player  {
 
 	public int id;
     public string name;
-    public float lat;
-    public float lng;
+    public double lat;
+    public double lng;
 
     public Player(int id, string name, double lat, double lng)
     {
         this.id = id;
         this.name = name;
-        this.lat = (float)lat;
-        this.lng = (float)lng;
+        this.lat = lat;
+        this.lng = lng;
     }
 }
